refactor(database): financeパッケージをdatabaseにリネーム
financeパッケージをdatabaseパッケージにリネームし、
プロジェクト全体のインポートを更新しました。

変更内容:
- src/finance/ を src/database/ にリネーム
- tests/finance/ を tests/database/ にリネーム
- pyproject.toml のパッケージ設定を更新
- 全ソースファイルのインポートを finance → database に更新
- ドキュメント・docstring内のコード例を更新

Closes #948
