+++ pyproject.toml
@@ -80,7 +80,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/finance", "src/market", "src/market_analysis", "src/rss", "src/quant", "src/factor", "src/strategy"]
+packages = ["src/database", "src/market", "src/market_analysis", "src/rss", "src/quant", "src/factor", "src/strategy"]
 
 [tool.ruff]
 line-length = 88
