+++ src/database/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility functions and configurations for the database package."""
+
+from database.utils.logging_config import get_logger
+
+__all__ = ["get_logger"]
