--- src/finance/__init__.py
@@ -1,6 +1,0 @@
-"""Finance - Cross-package database and utilities for finance project."""
-
-from finance.utils.logging_config import get_logger
-
-__version__ = "0.1.0"
-__all__ = ["__version__", "get_logger"]
