--- src/finance/utils/__init__.py
@@ -1,5 +1,0 @@
-"""Utility functions and configurations for the finance package."""
-
-from finance.utils.logging_config import get_logger
-
-__all__ = ["get_logger"]
