+++ src/market/utils/logging_config.py
@@ -8,7 +8,7 @@ from typing import Any
 
 from structlog import BoundLogger
 
-from finance.utils.logging_config import get_logger as _get_logger
+from database.utils.logging_config import get_logger as _get_logger
 
 
 def get_logger(name: str, **context: Any) -> BoundLogger:
