+++ src/market_analysis/utils/date_utils.py
@@ -27,7 +27,7 @@ from __future__ import annotations
 from datetime import date, datetime, timedelta
 from typing import Literal
 
-from finance.utils.logging_config import get_logger
+from database.utils.logging_config import get_logger
 
 logger = get_logger(__name__)
