+++ tests/rss/integration/test_feed_workflow.py
@@ -12,7 +12,7 @@ from pathlib import Path
 import pytest
 from pytest_httpserver import HTTPServer  # type: ignore[import-untyped]
 
-from finance.utils.logging_config import get_logger
+from database.utils.logging_config import get_logger
 from rss import (
     FeedFetcher,
     FeedManager,
